restore set compare hook
