+++ packages/signals/src/lib.rs
@@ -16,8 +16,8 @@ pub use read_only_signal::*;
 mod map;
 pub use map::*;
 
-// mod comparer;
-// pub use comparer::*;
+mod set_compare;
+pub use set_compare::*;
 
 mod memo;
 pub use memo::*;
